chore: fix implicit type casts

+++ bmp3.c
@@ -1,40 +1,40 @@
 /**
-* Copyright (c) 2022 Bosch Sensortec GmbH. All rights reserved.
-*
-* BSD-3-Clause
-*
-* Redistribution and use in source and binary forms, with or without
-* modification, are permitted provided that the following conditions are met:
-*
-* 1. Redistributions of source code must retain the above copyright
-*    notice, this list of conditions and the following disclaimer.
-*
-* 2. Redistributions in binary form must reproduce the above copyright
-*    notice, this list of conditions and the following disclaimer in the
-*    documentation and/or other materials provided with the distribution.
-*
-* 3. Neither the name of the copyright holder nor the names of its
-*    contributors may be used to endorse or promote products derived from
-*    this software without specific prior written permission.
-*
-* THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-* "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-* LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
-* FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
-* COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
-* INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-* (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-* SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
-* HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
-* STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING
-* IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-* POSSIBILITY OF SUCH DAMAGE.
-*
-* @file       bmp3.c
-* @date       2022-04-01
-* @version    v2.0.6
-*
-*/
+ * Copyright (c) 2022 Bosch Sensortec GmbH. All rights reserved.
+ *
+ * BSD-3-Clause
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of its
+ *    contributors may be used to endorse or promote products derived from
+ *    this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+ * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
+ * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING
+ * IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @file       bmp3.c
+ * @date       2022-04-01
+ * @version    v2.0.6
+ *
+ */
 
 /*! @file bmp3.c
  * @brief Sensor driver for BMP3 sensor */
@@ -1095,7 +1095,7 @@ int8_t bmp3_set_fifo_watermark(const struct bmp3_fifo_data *fifo,
 {
     int8_t rslt;
     uint8_t reg_data[2];
-    uint8_t reg_addr[2] = { BMP3_REG_FIFO_WM, BMP3_REG_FIFO_WM + 1 };
+    uint8_t reg_addr[2] = {BMP3_REG_FIFO_WM, BMP3_REG_FIFO_WM + 1};
     uint16_t watermark_len;
 
     if ((fifo != NULL) && (fifo_settings != NULL))
@@ -1153,7 +1153,7 @@ int8_t bmp3_extract_fifo_data(struct bmp3_data *data, struct bmp3_fifo_data *fif
     uint8_t header;
     uint8_t parsed_frames = 0;
     uint8_t t_p_frame;
-    struct bmp3_uncomp_data uncomp_data = { 0 };
+    struct bmp3_uncomp_data uncomp_data = {0};
 
     rslt = null_ptr_check(dev);
 
@@ -1431,8 +1431,8 @@ int8_t bmp3_get_sensor_data(uint8_t sensor_comp, struct bmp3_data *comp_data, st
 
     /* Array to store the pressure and temperature data read from
      * the sensor */
-    uint8_t reg_data[BMP3_LEN_P_T_DATA] = { 0 };
-    struct bmp3_uncomp_data uncomp_data = { 0 };
+    uint8_t reg_data[BMP3_LEN_P_T_DATA] = {0};
+    struct bmp3_uncomp_data uncomp_data = {0};
 
     if (comp_data != NULL)
     {
@@ -1539,39 +1539,39 @@ static uint8_t parse_fifo_data_frame(uint8_t header,
 
     switch (header)
     {
-        case BMP3_FIFO_TEMP_PRESS_FRAME:
-            unpack_temp_press_frame(byte_index, fifo->buffer, uncomp_data);
-            *parsed_frames = *parsed_frames + 1;
-            t_p_frame = BMP3_PRESS_TEMP;
-            break;
-        case BMP3_FIFO_TEMP_FRAME:
-            unpack_temp_frame(byte_index, fifo->buffer, uncomp_data);
-            *parsed_frames = *parsed_frames + 1;
-            t_p_frame = BMP3_TEMP;
-            break;
-        case BMP3_FIFO_PRESS_FRAME:
-            unpack_press_frame(byte_index, fifo->buffer, uncomp_data);
-            *parsed_frames = *parsed_frames + 1;
-            t_p_frame = BMP3_PRESS;
-            break;
-        case BMP3_FIFO_TIME_FRAME:
-            unpack_time_frame(byte_index, fifo->buffer, &fifo->sensor_time);
-            break;
-        case BMP3_FIFO_CONFIG_CHANGE:
-            fifo->config_change = 1;
-            *byte_index = *byte_index + 1;
-            break;
-        case BMP3_FIFO_ERROR_FRAME:
-            fifo->config_err = 1;
-            *byte_index = *byte_index + 1;
-            break;
-        case BMP3_FIFO_EMPTY_FRAME:
-            *byte_index = fifo->byte_count;
-            break;
-        default:
-            fifo->config_err = 1;
-            *byte_index = *byte_index + 1;
-            break;
+    case BMP3_FIFO_TEMP_PRESS_FRAME:
+        unpack_temp_press_frame(byte_index, fifo->buffer, uncomp_data);
+        *parsed_frames = *parsed_frames + 1;
+        t_p_frame = BMP3_PRESS_TEMP;
+        break;
+    case BMP3_FIFO_TEMP_FRAME:
+        unpack_temp_frame(byte_index, fifo->buffer, uncomp_data);
+        *parsed_frames = *parsed_frames + 1;
+        t_p_frame = BMP3_TEMP;
+        break;
+    case BMP3_FIFO_PRESS_FRAME:
+        unpack_press_frame(byte_index, fifo->buffer, uncomp_data);
+        *parsed_frames = *parsed_frames + 1;
+        t_p_frame = BMP3_PRESS;
+        break;
+    case BMP3_FIFO_TIME_FRAME:
+        unpack_time_frame(byte_index, fifo->buffer, &fifo->sensor_time);
+        break;
+    case BMP3_FIFO_CONFIG_CHANGE:
+        fifo->config_change = 1;
+        *byte_index = *byte_index + 1;
+        break;
+    case BMP3_FIFO_ERROR_FRAME:
+        fifo->config_err = 1;
+        *byte_index = *byte_index + 1;
+        break;
+    case BMP3_FIFO_EMPTY_FRAME:
+        *byte_index = fifo->byte_count;
+        break;
+    default:
+        fifo->config_err = 1;
+        *byte_index = *byte_index + 1;
+        break;
     }
 
     return t_p_frame;
@@ -1789,7 +1789,7 @@ static int8_t get_calib_data(struct bmp3_dev *dev)
     uint8_t reg_addr = BMP3_REG_CALIB_DATA;
 
     /* Array to store calibration data */
-    uint8_t calib_data[BMP3_LEN_CALIB_DATA] = { 0 };
+    uint8_t calib_data[BMP3_LEN_CALIB_DATA] = {0};
 
     /* Read the calibration data from the sensor */
     rslt = bmp3_get_regs(reg_addr, calib_data, BMP3_LEN_CALIB_DATA, dev);
@@ -1856,7 +1856,7 @@ static void parse_advance_settings(const uint8_t *reg_data, struct bmp3_adv_sett
  * @brief This internal API parse the power control(power mode, pressure enable
  * and temperature enable) settings and store in the device structure.
  */
-static void  parse_pwr_ctrl_settings(const uint8_t *reg_data, struct bmp3_settings *settings)
+static void parse_pwr_ctrl_settings(const uint8_t *reg_data, struct bmp3_settings *settings)
 {
     settings->op_mode = BMP3_GET_BITS(*reg_data, BMP3_OP_MODE);
     settings->press_en = BMP3_GET_BITS_POS_0(*reg_data, BMP3_PRESS_EN);
@@ -1867,7 +1867,7 @@ static void  parse_pwr_ctrl_settings(const uint8_t *reg_data, struct bmp3_settin
  * @brief This internal API parse the over sampling, ODR and filter
  * settings and store in the device structure.
  */
-static void  parse_odr_filter_settings(const uint8_t *reg_data, struct bmp3_odr_filter_settings *settings)
+static void parse_odr_filter_settings(const uint8_t *reg_data, struct bmp3_odr_filter_settings *settings)
 {
     uint8_t index = 0;
 
@@ -1930,7 +1930,7 @@ static int8_t set_odr_filter_settings(uint32_t desired_settings, struct bmp3_set
     int8_t rslt;
 
     /* No of registers to be configured is 3*/
-    uint8_t reg_addr[3] = { 0 };
+    uint8_t reg_addr[3] = {0};
 
     /* No of register data to be read is 4 */
     uint8_t reg_data[4];
@@ -1943,7 +1943,7 @@ static int8_t set_odr_filter_settings(uint32_t desired_settings, struct bmp3_set
         if (are_settings_changed((BMP3_SEL_PRESS_OS | BMP3_SEL_TEMP_OS), desired_settings))
         {
             /* Fill the over sampling register address and
-            * register data to be written in the sensor */
+             * register data to be written in the sensor */
             fill_osr_data(desired_settings, reg_addr, reg_data, &len, settings);
         }
 
@@ -2100,8 +2100,7 @@ static int8_t validate_osr_and_odr_settings(const struct bmp3_settings *settings
     /* Sampling period corresponding to ODR in microseconds  */
     uint32_t odr[18] = {
         5000, 10000, 20000, 40000, 80000, 160000, 320000, 640000, 1280000, 2560000, 5120000, 10240000, 20480000,
-        40960000, 81920000, 163840000, 327680000, 655360000
-    };
+        40960000, 81920000, 163840000, 327680000, 655360000};
 
     if (settings->press_en)
     {
@@ -2445,18 +2444,18 @@ static int8_t compensate_temperature(double *temperature,
     /* Update the compensated temperature in calib structure since this is
      * needed for pressure calculation */
     calib_data->quantized_calib_data.t_lin = partial_data2 + (partial_data1 * partial_data1) *
-                                             calib_data->quantized_calib_data.par_t3;
+                                                                 calib_data->quantized_calib_data.par_t3;
 
     /* Returns compensated temperature */
-    if (calib_data->quantized_calib_data.t_lin < BMP3_MIN_TEMP_DOUBLE)
+    if (calib_data->quantized_calib_data.t_lin < (double)BMP3_MIN_TEMP_DOUBLE)
     {
-        calib_data->quantized_calib_data.t_lin = BMP3_MIN_TEMP_DOUBLE;
+        calib_data->quantized_calib_data.t_lin = (double)BMP3_MIN_TEMP_DOUBLE;
         rslt = BMP3_W_MIN_TEMP;
     }
 
-    if (calib_data->quantized_calib_data.t_lin > BMP3_MAX_TEMP_DOUBLE)
+    if (calib_data->quantized_calib_data.t_lin > (double)BMP3_MAX_TEMP_DOUBLE)
     {
-        calib_data->quantized_calib_data.t_lin = BMP3_MAX_TEMP_DOUBLE;
+        calib_data->quantized_calib_data.t_lin = (double)BMP3_MAX_TEMP_DOUBLE;
         rslt = BMP3_W_MAX_TEMP;
     }
 
@@ -2489,27 +2488,27 @@ static int8_t compensate_pressure(double *pressure,
     double partial_out2;
 
     partial_data1 = quantized_calib_data->par_p6 * quantized_calib_data->t_lin;
-    partial_data2 = quantized_calib_data->par_p7 * pow_bmp3(quantized_calib_data->t_lin, 2);
-    partial_data3 = quantized_calib_data->par_p8 * pow_bmp3(quantized_calib_data->t_lin, 3);
+    partial_data2 = quantized_calib_data->par_p7 * (double)pow_bmp3(quantized_calib_data->t_lin, 2);
+    partial_data3 = quantized_calib_data->par_p8 * (double)pow_bmp3(quantized_calib_data->t_lin, 3);
     partial_out1 = quantized_calib_data->par_p5 + partial_data1 + partial_data2 + partial_data3;
     partial_data1 = quantized_calib_data->par_p2 * quantized_calib_data->t_lin;
-    partial_data2 = quantized_calib_data->par_p3 * pow_bmp3(quantized_calib_data->t_lin, 2);
-    partial_data3 = quantized_calib_data->par_p4 * pow_bmp3(quantized_calib_data->t_lin, 3);
+    partial_data2 = quantized_calib_data->par_p3 * (double)pow_bmp3(quantized_calib_data->t_lin, 2);
+    partial_data3 = quantized_calib_data->par_p4 * (double)pow_bmp3(quantized_calib_data->t_lin, 3);
     partial_out2 = uncomp_data->pressure *
                    (quantized_calib_data->par_p1 + partial_data1 + partial_data2 + partial_data3);
     partial_data1 = pow_bmp3((double)uncomp_data->pressure, 2);
     partial_data2 = quantized_calib_data->par_p9 + quantized_calib_data->par_p10 * quantized_calib_data->t_lin;
     partial_data3 = partial_data1 * partial_data2;
-    partial_data4 = partial_data3 + pow_bmp3((double)uncomp_data->pressure, 3) * quantized_calib_data->par_p11;
+    partial_data4 = partial_data3 + (double)pow_bmp3((double)uncomp_data->pressure, 3) * quantized_calib_data->par_p11;
     comp_press = partial_out1 + partial_out2 + partial_data4;
 
-    if (comp_press < BMP3_MIN_PRES_DOUBLE)
+    if (comp_press < (double)BMP3_MIN_PRES_DOUBLE)
     {
         comp_press = BMP3_MIN_PRES_DOUBLE;
         rslt = BMP3_W_MIN_PRES;
     }
 
-    if (comp_press > BMP3_MAX_PRES_DOUBLE)
+    if (comp_press > (double)BMP3_MAX_PRES_DOUBLE)
     {
         comp_press = BMP3_MAX_PRES_DOUBLE;
         rslt = BMP3_W_MAX_PRES;
@@ -2530,7 +2529,7 @@ static float pow_bmp3(double base, uint8_t power)
 
     while (power != 0)
     {
-        pow_output = (float) base * pow_output;
+        pow_output = (float)base * pow_output;
         power--;
     }
